Run isValidElement tests against minified source
