+++ test/browser/isValidElement.test.js
@@ -1,0 +1,4 @@
+import { createElement, isValidElement, Component } from 'preact';
+import { isValidElementTests } from '../shared/isValidElementTests';
+
+isValidElementTests(expect, isValidElement, createElement, Component);
